Move schema and queries to their owners; drop tauri-plugin-sql (phase 2+3)
The frontend talks to the generic Rust layer through DatabaseAccess,
whose transaction(callback) is replaced by batch(statements): a list of
statements that runs atomically in one call. Ids are resolved by
subselect inside the statements, so nothing needs a round trip mid-
transaction and nothing can be left half-committed.

Schema generation 2 lives with its owners as a single init migration
each — app (history), workspace feature, side-menu ui-state feature —
and reaches Rust through the existing databaseMigrations contribution
slot. The twelve generation-1 files are gone; the database is a new
file (cogno-v2.db), the old one stays untouched for the import step.

Repositories are rewritten against the new schema:
- HistoryRepository builds one batch per write, uses trigram FTS for
  fragments of three characters or more, derives transition totals and
  slot statistics with SUM/GROUP BY, and forgets a command's log rows
  along with its ranking row. Command duration and return code are now
  stored.
- WorkspaceRepository loads everything in three queries instead of
  1 + 2N and writes each change as one batch.
- SideMenuStateRepository upserts instead of updating a seeded row.

TerminalHistoryPersistenceService receives DatabaseAccess through DI;
its public API is unchanged, so the autocomplete suggestors are
untouched. tauri-plugin-sql and its capability are removed, which also
lifts the libsqlite3-sys pin: rusqlite moves to 0.40.

+++ src-tauri/src/db/mod.rs
@@ -454,14 +454,14 @@ mod tests {
                         tx.execute("INSERT INTO big (payload) VALUES (?1)", [&payload])?;
                     }
                     tx.commit()?;
-                    let page_size: u64 =
+                    let page_size: i64 =
                         conn.query_row("PRAGMA page_size", [], |r| r.get(0))?;
-                    let root: u64 = conn.query_row(
+                    let root: i64 = conn.query_row(
                         "SELECT rootpage FROM sqlite_master WHERE name = 'big'",
                         [],
                         |r| r.get(0),
                     )?;
-                    Ok((page_size, root))
+                    Ok((page_size as u64, root as u64))
                 })
                 .unwrap();
             db.close();
